+++ examples/ed25519_dalek.rs
@@ -1,17 +1,3 @@
-// Copyright 2018 Alex Ostrovski
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 //! Example demonstrating how to use the crate with external types which don't implement
 //! any "useful" traits (e.g., `AsRef<[u8]>` or `FromHex`).
 //!
@@ -22,12 +8,11 @@
 extern crate alloc;
 
 use ed25519::{PublicKey, SecretKey};
-use rand::thread_rng;
 use serde_derive::*;
 
 use alloc::{
     borrow::{Cow, ToOwned},
-    string::{String, ToString},
+    string::String,
 };
 
 use hex_buffer_serde::Hex;
@@ -35,12 +20,14 @@ use hex_buffer_serde::Hex;
 struct PublicKeyHex(());
 
 impl Hex<PublicKey> for PublicKeyHex {
+    type Error = ed25519::SignatureError;
+
     fn create_bytes(value: &PublicKey) -> Cow<'_, [u8]> {
         Cow::Borrowed(&*value.as_bytes())
     }
 
-    fn from_bytes(bytes: &[u8]) -> Result<PublicKey, String> {
-        PublicKey::from_bytes(bytes).map_err(|e| e.to_string())
+    fn from_bytes(bytes: &[u8]) -> Result<PublicKey, Self::Error> {
+        PublicKey::from_bytes(bytes)
     }
 }
 
@@ -55,7 +42,9 @@ struct SomeData {
 }
 
 fn main() {
-    let secret_key = SecretKey::generate(&mut thread_rng());
+    let secret_key =
+        hex::decode("9e55d1e1aa1f455b8baad9fdf975503655f8b359d542fa7e4ce84106d625b352").unwrap();
+    let secret_key = SecretKey::from_bytes(&secret_key).unwrap();
     let public_key: PublicKey = (&secret_key).into();
 
     let key_hex = hex::encode(public_key.as_bytes());