+++ src/lib.rs
@@ -1,17 +1,3 @@
-// Copyright 2020 Alex Ostrovski
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 //! Serializing byte buffers as hex strings with `serde`.
 //!
 //! # Problem
@@ -31,14 +17,17 @@
 //! designated specifically for this purpose. The implementor can then be used
 //! for (de)serialization with the help of the `#[serde(with)]` attribute.
 //!
-//! [`sodiumoxide`]: https://crates.io/crates/sodiumoxide
+//! [`ConstHex`] is an analogue of [`Hex`] that can be used if the serialized buffer has
+//! constant length known in compile time.
 //!
 //! # Crate Features
 //!
-//! - `std` (enabled by default): Enables types from the Rust standard library. Switching
-//!   this feature off can be used for constrained environments, such as WASM. Note that
-//!   the crate still requires an allocator (that is, the `alloc` crate) even
-//!   if the `std` feature is disabled.
+//! - `alloc` (enabled by default). Enables types that depend on the `alloc` crate:
+//!   [`Hex`] and [`HexForm`].
+//! - `const_len` (disabled by default). Enables types that depend on const generics:
+//!   [`ConstHex`] and [`ConstHexForm`]. Requires Rust 1.51 or newer.
+//!
+//! [`sodiumoxide`]: https://crates.io/crates/sodiumoxide
 //!
 //! # Examples
 //!
@@ -61,17 +50,19 @@
 //!
 //! // We define in our crate:
 //! use hex_buffer_serde::Hex;
-//! use serde_derive::*;
+//! use serde_derive::{Deserialize, Serialize};
 //!
 //! # use std::borrow::Cow;
-//! enum BufferHex {} // a single-purpose type for use in `#[serde(with)]`
+//! struct BufferHex; // a single-purpose type for use in `#[serde(with)]`
 //! impl Hex<Buffer> for BufferHex {
+//!     type Error = &'static str;
+//!
 //!     fn create_bytes(buffer: &Buffer) -> Cow<[u8]> {
 //!         buffer.as_ref().into()
 //!     }
 //!
-//!     fn from_bytes(bytes: &[u8]) -> Result<Buffer, String> {
-//!         Buffer::from_slice(bytes).ok_or_else(|| "invalid byte length".to_owned())
+//!     fn from_bytes(bytes: &[u8]) -> Result<Buffer, Self::Error> {
+//!         Buffer::from_slice(bytes).ok_or_else(|| "invalid byte length")
 //!     }
 //! }
 //!
@@ -85,7 +76,7 @@
 //! # fn main() {}
 //! ```
 //!
-//! # Use with internal types
+//! ## Use with internal types
 //!
 //! The crate could still be useful if you have control over the serialized buffer type.
 //! `Hex<T>` has a blanket implementation for types `T` satisfying certain constraints:
@@ -127,336 +118,31 @@
 //! ```
 
 #![no_std]
+// Documentation settings.
+#![cfg_attr(docsrs, feature(doc_cfg))]
 #![doc(html_root_url = "https://docs.rs/hex-buffer-serde/0.2.2")]
+// Linter settings.
 #![warn(missing_docs, missing_debug_implementations)]
 #![warn(clippy::all, clippy::pedantic)]
 #![allow(clippy::missing_errors_doc, clippy::must_use_candidate)]
 
+#[cfg(any(test, feature = "alloc"))]
 extern crate alloc;
 
-use serde::{de::Visitor, Deserializer, Serializer};
+#[cfg(feature = "const_len")]
+mod const_len;
+#[cfg(feature = "const_len")]
+pub use self::const_len::{ConstHex, ConstHexForm};
+#[cfg(feature = "alloc")]
+mod var_len;
+#[cfg(feature = "alloc")]
+pub use self::var_len::{Hex, HexForm};
 
-use alloc::{
-    borrow::Cow,
-    string::{String, ToString},
-    vec::Vec,
-};
-use core::{convert::TryFrom, fmt, marker::PhantomData};
+#[cfg(not(any(feature = "const_len", feature = "alloc")))]
+compile_error!(
+    "At least one of `const_len` and `alloc` features must be enabled; \
+     the crate is useless otherwise"
+);
 
-/// Provides hex-encoded (de)serialization for `serde`.
-///
-/// Note that the trait is automatically implemented for types that
-/// implement `AsRef<[u8]>` and `TryFrom<&[u8]>`.
-pub trait Hex<T> {
-    /// Converts the value into bytes. This is used for serialization.
-    ///
-    /// The returned buffer can be either borrowed from the type, or created by the method.
-    fn create_bytes(value: &T) -> Cow<'_, [u8]>;
-
-    /// Creates a value from the byte slice.
-    ///
-    /// # Errors
-    ///
-    /// If this method fails, it should return a human-readable error description conforming
-    /// to `serde` conventions (no upper-casing of the first letter, no punctuation at the end).
-    fn from_bytes(bytes: &[u8]) -> Result<T, String>;
-
-    /// Serializes the value for `serde`. This method is not meant to be overridden.
-    ///
-    /// The serialization is a lower-case hex string
-    /// for [human-readable][hr] serializers (e.g., JSON or TOML), and the original bytes
-    /// returned by [`Self::create_bytes()`] for non-human-readable ones.
-    ///
-    /// [hr]: serde::Serializer::is_human_readable()
-    /// [`create_bytes`]: #tymethod.create_bytes
-    fn serialize<S: Serializer>(value: &T, serializer: S) -> Result<S::Ok, S::Error> {
-        let value = Self::create_bytes(value);
-        if serializer.is_human_readable() {
-            serializer.serialize_str(&hex::encode(value))
-        } else {
-            serializer.serialize_bytes(value.as_ref())
-        }
-    }
-
-    /// Deserializes a value using `serde`. This method is not meant to be overridden.
-    ///
-    /// If the deserializer is [human-readable][hr] (e.g., JSON or TOML), this method
-    /// expects a hex-encoded string. Otherwise, the method expects a byte array.
-    ///
-    /// [hr]: serde::Serializer::is_human_readable()
-    fn deserialize<'de, D>(deserializer: D) -> Result<T, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        use serde::de::Error as DeError;
-
-        struct HexVisitor;
-
-        impl<'de> Visitor<'de> for HexVisitor {
-            type Value = Vec<u8>;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-                formatter.write_str("hex-encoded byte array")
-            }
-
-            fn visit_str<E: DeError>(self, value: &str) -> Result<Self::Value, E> {
-                hex::decode(value).map_err(E::custom)
-            }
-
-            // See the `deserializing_flattened_field` test for an example why this is needed.
-            fn visit_bytes<E: DeError>(self, value: &[u8]) -> Result<Self::Value, E> {
-                Ok(value.to_vec())
-            }
-        }
-
-        struct BytesVisitor;
-
-        impl<'de> Visitor<'de> for BytesVisitor {
-            type Value = Vec<u8>;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-                formatter.write_str("byte array")
-            }
-
-            fn visit_bytes<E: DeError>(self, value: &[u8]) -> Result<Self::Value, E> {
-                Ok(value.to_vec())
-            }
-        }
-
-        let maybe_bytes = if deserializer.is_human_readable() {
-            deserializer.deserialize_str(HexVisitor)
-        } else {
-            deserializer.deserialize_bytes(BytesVisitor)
-        };
-        maybe_bytes.and_then(|bytes| Self::from_bytes(&bytes).map_err(D::Error::custom))
-    }
-}
-
-/// A dummy container for use inside `#[serde(with)]` attribute.
-///
-/// # Why a separate container?
-///
-/// We need a separate type (instead of just using `impl<T> Hex<T> for T`)
-/// both for code clarity and because otherwise invocations within generated `serde` code
-/// would be ambiguous for types implementing `Serialize` / `Deserialize`.
-#[derive(Debug)]
-pub struct HexForm<T>(PhantomData<T>);
-
-impl<T> Hex<T> for HexForm<T>
-where
-    T: AsRef<[u8]> + for<'a> TryFrom<&'a [u8]>,
-    for<'a> <T as TryFrom<&'a [u8]>>::Error: ToString,
-{
-    fn create_bytes(buffer: &T) -> Cow<'_, [u8]> {
-        Cow::Borrowed(buffer.as_ref())
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Result<T, String> {
-        T::try_from(bytes).map_err(|e| e.to_string())
-    }
-}
-
-#[cfg(test)]
-#[allow(renamed_and_removed_lints, clippy::unknown_clippy_lints)]
-// ^ `map_err_ignore` is newer than MSRV, and `clippy::unknown_clippy_lints` is removed
-// since Rust 1.51.
-mod tests {
-    use serde_derive::{Deserialize, Serialize};
-    use serde_json::json;
-
-    use super::*;
-    use alloc::{borrow::ToOwned, vec};
-    use core::array::TryFromSliceError;
-
-    #[test]
-    fn internal_type() {
-        pub struct Buffer([u8; 8]);
-
-        impl AsRef<[u8]> for Buffer {
-            fn as_ref(&self) -> &[u8] {
-                &self.0
-            }
-        }
-
-        impl TryFrom<&[u8]> for Buffer {
-            type Error = String;
-
-            #[allow(clippy::map_err_ignore)]
-            fn try_from(slice: &[u8]) -> Result<Self, Self::Error> {
-                <[u8; 8]>::try_from(slice)
-                    .map(Buffer)
-                    .map_err(|_| "!".to_owned())
-            }
-        }
-
-        #[derive(Serialize, Deserialize)]
-        struct Test {
-            #[serde(with = "HexForm::<Buffer>")]
-            buffer: Buffer,
-            other_field: String,
-        }
-
-        let json = json!({ "buffer": "0001020304050607", "other_field": "abc" });
-        let value: Test = serde_json::from_value(json.clone()).unwrap();
-        assert!(value
-            .buffer
-            .0
-            .iter()
-            .enumerate()
-            .all(|(i, &byte)| i == usize::from(byte)));
-
-        let json_copy = serde_json::to_value(&value).unwrap();
-        assert_eq!(json, json_copy);
-    }
-
-    #[test]
-    fn internal_type_with_derived_serde_code() {
-        #[derive(Serialize, Deserialize)]
-        pub struct Buffer([u8; 8]);
-
-        impl AsRef<[u8]> for Buffer {
-            fn as_ref(&self) -> &[u8] {
-                &self.0
-            }
-        }
-
-        impl TryFrom<&[u8]> for Buffer {
-            type Error = TryFromSliceError;
-
-            fn try_from(slice: &[u8]) -> Result<Self, Self::Error> {
-                <[u8; 8]>::try_from(slice).map(Buffer)
-            }
-        }
-
-        // here, a hex form should be used.
-        #[derive(Serialize, Deserialize)]
-        struct HexTest {
-            #[serde(with = "HexForm::<Buffer>")]
-            buffer: Buffer,
-            other_field: String,
-        }
-
-        // ...and here, we may use original `serde` code.
-        #[derive(Serialize, Deserialize)]
-        struct OriginalTest {
-            buffer: Buffer,
-            other_field: String,
-        }
-
-        let test = HexTest {
-            buffer: Buffer([1; 8]),
-            other_field: "a".to_owned(),
-        };
-        assert_eq!(
-            serde_json::to_value(test).unwrap(),
-            json!({
-                "buffer": "0101010101010101",
-                "other_field": "a",
-            })
-        );
-
-        let test = OriginalTest {
-            buffer: Buffer([1; 8]),
-            other_field: "a".to_owned(),
-        };
-        assert_eq!(
-            serde_json::to_value(test).unwrap(),
-            json!({
-                "buffer": [1, 1, 1, 1, 1, 1, 1, 1],
-                "other_field": "a",
-            })
-        );
-    }
-
-    #[test]
-    fn external_type() {
-        #[derive(Debug, PartialEq, Eq)]
-        pub struct Buffer([u8; 8]);
-
-        struct BufferHex(());
-
-        impl Hex<Buffer> for BufferHex {
-            fn create_bytes(buffer: &Buffer) -> Cow<'_, [u8]> {
-                Cow::Borrowed(&buffer.0)
-            }
-
-            fn from_bytes(bytes: &[u8]) -> Result<Buffer, String> {
-                if bytes.len() == 8 {
-                    let mut inner = [0; 8];
-                    inner.copy_from_slice(bytes);
-                    Ok(Buffer(inner))
-                } else {
-                    Err("invalid buffer length".to_owned())
-                }
-            }
-        }
-
-        #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-        struct Test {
-            #[serde(with = "BufferHex")]
-            buffer: Buffer,
-            other_field: String,
-        }
-
-        let json = json!({ "buffer": "0001020304050607", "other_field": "abc" });
-        let value: Test = serde_json::from_value(json.clone()).unwrap();
-        assert!(value
-            .buffer
-            .0
-            .iter()
-            .enumerate()
-            .all(|(i, &byte)| i == usize::from(byte)));
-
-        let json_copy = serde_json::to_value(&value).unwrap();
-        assert_eq!(json, json_copy);
-
-        // Test binary / non-human readable format.
-        let buffer = bincode::serialize(&value).unwrap();
-        // Conversion to hex is needed to be able to search for a pattern.
-        let buffer_hex = hex::encode(&buffer);
-        // Check that the buffer is stored in the serialization compactly,
-        // as original bytes.
-        let needle = "0001020304050607";
-        assert!(buffer_hex.contains(needle));
-
-        let value_copy: Test = bincode::deserialize(&buffer).unwrap();
-        assert_eq!(value_copy, value);
-    }
-
-    #[test]
-    fn deserializing_flattened_field() {
-        // The fields in the flattened structure are somehow read with
-        // a human-readable `Deserializer`, even if the original `Deserializer`
-        // is not human-readable.
-        #[derive(Debug, PartialEq, Serialize, Deserialize)]
-        struct Inner {
-            #[serde(with = "HexForm")]
-            x: Vec<u8>,
-            #[serde(with = "HexForm")]
-            y: [u8; 16],
-        }
-
-        #[derive(Debug, PartialEq, Serialize, Deserialize)]
-        struct Outer {
-            #[serde(flatten)]
-            inner: Inner,
-            z: String,
-        }
-
-        let value = Outer {
-            inner: Inner {
-                x: vec![1; 8],
-                y: [0; 16],
-            },
-            z: "test".to_owned(),
-        };
-
-        let bytes = serde_cbor::to_vec(&value).unwrap();
-        let bytes_hex = hex::encode(&bytes);
-        // Check that byte buffers are stored in the binary form.
-        assert!(bytes_hex.contains(&"01".repeat(8)));
-        assert!(bytes_hex.contains(&"00".repeat(16)));
-        let value_copy = serde_cbor::from_slice(&bytes).unwrap();
-        assert_eq!(value, value_copy);
-    }
-}
+#[cfg(doctest)]
+doc_comment::doctest!("../README.md");